# tracing: manually redact string in spanInner.Recordf