@dhartunian 2022-01-31
I experimented with adding this shortcut in the `redact` repository but did not see useful results.

```
david@mim ~/g/s/g/c/redact (add-sprintf-shortcut)> go test -bench=. -benchmem
goos: linux
goarch: amd64
pkg: github.com/cockroachdb/redact
cpu: Intel(R) Core(TM) i5-4670 CPU @ 3.40GHz
BenchmarkSprintf-4       	 1490469	       813.5 ns/op	      88 B/op	       2 allocs/op
BenchmarkSprintfFast-4   	 1753380	       692.1 ns/op	     112 B/op	       3 allocs/op
PASS
ok  	github.com/cockroachdb/redact	5.160s
```

Attempting to repeat the experiment above led to the opposite results:
```
name                                            old time/op    new time/op    delta
Eventf_WithVerboseTraceSpan/redactable=false-4    3.56µs ± 1%    3.56µs ± 2%     ~     (p=0.670 n=10+10)
Eventf_WithVerboseTraceSpan/redactable=true-4     4.57µs ± 1%    4.58µs ± 1%     ~     (p=0.305 n=10+9)

name                                            old alloc/op   new alloc/op   delta
Eventf_WithVerboseTraceSpan/redactable=false-4      809B ± 0%     1112B ± 0%  +37.45%  (p=0.000 n=9+8)
Eventf_WithVerboseTraceSpan/redactable=true-4       977B ± 0%      977B ± 0%     ~     (all equal)

name                                            old allocs/op  new allocs/op  delta
Eventf_WithVerboseTraceSpan/redactable=false-4      18.0 ± 0%      21.0 ± 0%  +16.67%  (p=0.000 n=10+10)
Eventf_WithVerboseTraceSpan/redactable=true-4       19.0 ± 0%      19.0 ± 0%     ~     (all equal)
```

These new stats are run on Linux so perhaps there's something different here but I'm closing this either way since there's no clear path forward. It's possible that in all the redaction work previously, I mixed up the ordering of some of the benchmark runs.